[refactor] refactor full-stack architecture
