+++ next.config.ts
@@ -1,7 +1,7 @@
 import type { NextConfig } from 'next';
 import createNextIntlPlugin from 'next-intl/plugin';
 
-const withNextIntl = createNextIntlPlugin();
+const withNextIntl = createNextIntlPlugin('./src/client/i18n/request.ts');
 const nextConfig: NextConfig = {};
 
 export default withNextIntl(nextConfig);
