--- src/app/[locale]/_components/index.ts
@@ -1,3 +1,0 @@
-export { default as Header } from './header';
-export { default as SignInToast } from './signInToast';
-export { default as FeatureCard } from './featureCard';
