+++ src/app/[locale]/auth/callback/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { createServerClient } from '@/lib/auth/server';
+import { createServerClient } from '@/server/core/auth';
 
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url);
