--- src/app/[locale]/workspace/_components/index.ts
@@ -1,2 +1,0 @@
-export { default as Sidebar } from './sidebar';
-export { default as ExpandableArea } from './sidebar/expandableArea';
