+++ src/app/[locale]/workspace/layout.tsx
@@ -1,4 +1,5 @@
-import { Sidebar, ExpandableArea } from './_components';
+import Sidebar from '@/client/components/domain/sidebar';
+import ExpandableArea from '@/client/components/domain/sidebar/expandableArea';
 
 export default function WorkspaceLayout({ children }: { children: React.ReactNode }) {
   return (
