+++ src/client/components/domain/header/authButton.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import { useRouter, usePathname } from '@/i18n/routing';
-import { useUser } from '@/hooks';
-import { RoundButton } from '@/components/ui/button';
-import { rootPaths } from '@/constants';
+import { useRouter, usePathname } from '@/client/i18n/routing';
+import { useUser } from '@/client/hooks';
+import { RoundButton } from '@/client/components/ui/button';
+import { rootPaths } from '@/shared/constants';
 import { useTranslations } from 'next-intl';
 
 export default function AuthButton() {
