+++ src/client/components/domain/pages/onboarding/index.ts
@@ -1,0 +1,1 @@
+export { default as FeatureCard } from './featureCard';
