+++ src/client/components/domain/sidebar/menuItems.tsx
