+++ src/client/components/domain/toast/index.ts
@@ -1,0 +1,1 @@
+export { default as SignInToast } from './signInToast';
