+++ src/client/components/domain/toast/signInToast.tsx
@@ -3,9 +3,9 @@
 import { useEffect } from 'react';
 import { useSearchParams } from 'next/navigation';
 import { toast } from 'react-toastify';
-import { toastConfig } from '@/constants/toastConfig';
+import { toastConfig } from '@/client/constants/toastConfig';
 import { useTranslations } from 'next-intl';
-import { useRouter } from '@/i18n/routing';
+import { useRouter } from '@/client/i18n/routing';
 
 export default function SignInToast() {
   const searchParams = useSearchParams();
