+++ src/client/messages/en/common/components.json
