+++ src/client/messages/ja/common/components.json
