+++ src/client/messages/ja/metadata.json
