+++ src/client/messages/ja/pages/onboarding.json
