+++ src/client/messages/ja/pages/workspace.json
