+++ src/client/messages/ko/common/components.json
